+++ src/main/java/eu/stiekema/jeroen/adventofcode2019/common/StringParseUtil.java
@@ -6,11 +6,11 @@ import java.util.List;
 import java.util.StringTokenizer;
 
 public class StringParseUtil {
-    public static List<Integer> getCodes(String commaSeparatedValues, String delimiter) {
+    public static List<Long> getCodes(String commaSeparatedValues, String delimiter) {
         StringTokenizer st = new StringTokenizer(commaSeparatedValues, delimiter);
-        List<Integer> codes = new ArrayList<>();
+        List<Long> codes = new ArrayList<>();
         while (st.hasMoreTokens()) {
-            codes.add(Integer.valueOf(st.nextToken()));
+            codes.add(Long.valueOf(st.nextToken()));
         }
         return Collections.unmodifiableList(codes);
     }
