+++ src/main/java/eu/stiekema/jeroen/adventofcode2019/day5/Main.java
@@ -12,7 +12,7 @@ public class Main {
     public static final int THERMAL_RADIATOR_ID = 5;
 
     public static void main(String[] args) throws IOException {
-        List<Integer> codes = FileParseUtil.getCodes("day5.txt", ",");
+        List<Long> codes = FileParseUtil.getCodes("day5.txt", ",");
 
         IntcodeComputer intcodeComputer = IntcodeComputer.newInstance(codes);
         intcodeComputer.addInput(AIRCONDITIONER_UNIT_ID);
