+++ src/test/java/eu/stiekema/jeroen/adventofcode2019/intcode/IntcodeComputerTest.java
@@ -11,18 +11,18 @@ public class IntcodeComputerTest {
     public void interpretAndReturnDiagnosticCode() {
         String program = "3,3,1107,-1,8,3,4,3,99";
 
-        assertThat(executeWithInput(program, 1)).isEqualTo(1);
-        assertThat(executeWithInput(program, 2)).isEqualTo(1);
-        assertThat(executeWithInput(program, 3)).isEqualTo(1);
-        assertThat(executeWithInput(program, 4)).isEqualTo(1);
-        assertThat(executeWithInput(program, 5)).isEqualTo(1);
-        assertThat(executeWithInput(program, 6)).isEqualTo(1);
-        assertThat(executeWithInput(program, 7)).isEqualTo(1);
-        assertThat(executeWithInput(program, 8)).isEqualTo(0);
+        assertThat(executeWithInput(program, 1)).isEqualTo(1L);
+        assertThat(executeWithInput(program, 2)).isEqualTo(1L);
+        assertThat(executeWithInput(program, 3)).isEqualTo(1L);
+        assertThat(executeWithInput(program, 4)).isEqualTo(1L);
+        assertThat(executeWithInput(program, 5)).isEqualTo(1L);
+        assertThat(executeWithInput(program, 6)).isEqualTo(1L);
+        assertThat(executeWithInput(program, 7)).isEqualTo(1L);
+        assertThat(executeWithInput(program, 8)).isEqualTo(0L);
 
     }
 
-    private int executeWithInput(String program, int input) {
+    private long executeWithInput(String program, int input) {
         IntcodeComputer intcodeComputer = IntcodeComputer.newInstance(StringParseUtil.getCodes(program, ","));
         intcodeComputer.addInput(input);
         return intcodeComputer.executeDiagnostic();
